Windows

+++ egui/code/components/treenode/etreenode.cpp
@@ -717,7 +717,7 @@ void eTreeNode::draw_value(
 
         ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, zero_pad);
         if (value_w < 0) ImGui::SetNextItemWidth(-FLT_MIN);
-        ImGui::InputText(label, m_edit_buf.ptr(), m_edit_buf.sz(), eflags);
+        ImGui::InputText(label, m_edit_buf.ptr(), (size_t)m_edit_buf.sz(), eflags);
         if ((!ImGui::IsItemActive() || ImGui::IsItemDeactivatedAfterEdit()) && m_prev_edit_value)
         {
             eVariable value, nice_value;
